Rename folder Account to Resources and depreciated the Account class
