+++ src/tagoio_sdk/modules/Resources/Billing.py
@@ -1,10 +1,7 @@
 from tagoio_sdk.common.tagoio_module import TagoIOModule
-from tagoio_sdk.modules.Account.Billing_Type import (
-    BillingEditResourceAllocation,
-    BillingEditSubscription,
-    BillingPrices,
-    BillingSubscription,
-)
+from tagoio_sdk.modules.Resources.Billing_Type import (
+    BillingEditResourceAllocation, BillingEditSubscription, BillingPrices,
+    BillingSubscription)
 
 
 class Billing(TagoIOModule):
