+++ src/tagoio_sdk/modules/Resources/Buckets_Type.py
