+++ src/tagoio_sdk/modules/Resources/Dashboards_Type.py
@@ -1,13 +1,9 @@
 from datetime import datetime
 from typing import Literal, Optional, TypedDict, Union
 
-from tagoio_sdk.common.Common_Type import (
-    ExpireTimeOption,
-    GenericID,
-    GenericToken,
-    TagsObj,
-)
-from tagoio_sdk.modules.Account.Buckets_Type import BucketDeviceInfo
+from tagoio_sdk.common.Common_Type import (ExpireTimeOption, GenericID,
+                                           GenericToken, TagsObj)
+from tagoio_sdk.modules.Resources.Buckets_Type import BucketDeviceInfo
 
 
 class Arrangement(TypedDict):
