+++ src/tagoio_sdk/modules/Resources/Device_Type.py
@@ -1,14 +1,10 @@
 from datetime import datetime
-from typing import Union, Literal, Optional, TypedDict
+from typing import Literal, Optional, TypedDict, Union
 
-from tagoio_sdk.common.Common_Type import (
-    ExpireTimeOption,
-    GenericID,
-    GenericToken,
-    PermissionOption,
-    TagsObj,
-)
-from tagoio_sdk.modules.Account.Buckets_Type import DataStorageType
+from tagoio_sdk.common.Common_Type import (ExpireTimeOption, GenericID,
+                                           GenericToken, PermissionOption,
+                                           TagsObj)
+from tagoio_sdk.modules.Resources.Buckets_Type import DataStorageType
 
 
 class bucket(TypedDict):
