+++ src/tagoio_sdk/modules/Resources/Devices.py
@@ -1,24 +1,19 @@
 from typing import Optional
-from tagoio_sdk.common.Common_Type import (
-    Data,
-    GenericID,
-    GenericToken,
-    TokenCreateResponse,
-)
+
+from tagoio_sdk.common.Common_Type import (Data, GenericID, GenericToken,
+                                           TokenCreateResponse)
 from tagoio_sdk.common.tagoio_module import TagoIOModule
-from tagoio_sdk.modules.Account.Device_Type import (
-    ConfigurationParams,
-    DeviceCreateInfo,
-    DeviceCreateResponse,
-    DeviceEditInfo,
-    DeviceListItem,
-    DeviceQuery,
-    DeviceTokenDataList,
-    ListDeviceTokenQuery,
-    TokenData,
-)
 from tagoio_sdk.modules.Device.Device_Type import DataQuery, DeviceInfo
-from tagoio_sdk.modules.Utils.dateParser import dateParserList, dateParser
+from tagoio_sdk.modules.Resources.Device_Type import (ConfigurationParams,
+                                                      DeviceCreateInfo,
+                                                      DeviceCreateResponse,
+                                                      DeviceEditInfo,
+                                                      DeviceListItem,
+                                                      DeviceQuery,
+                                                      DeviceTokenDataList,
+                                                      ListDeviceTokenQuery,
+                                                      TokenData)
+from tagoio_sdk.modules.Utils.dateParser import dateParser, dateParserList
 
 
 class Devices(TagoIOModule):
