+++ src/tagoio_sdk/modules/Resources/Run.py
@@ -1,22 +1,16 @@
 from typing import Optional, TypedDict
+
 from tagoio_sdk.common.Common_Type import GenericID, Query
 from tagoio_sdk.common.tagoio_module import TagoIOModule
-from tagoio_sdk.modules.Account.Notification_Type import (
-    NotificationCreate,
-    NotificationInfo,
-    NotificationcreateReturn,
-)
-from tagoio_sdk.modules.Account.Run_Type import (
-    CustomDomainCreate,
-    CustomDomainInfo,
-    LoginAsUserOptions,
-    LoginResponse,
-    RunInfo,
-    RunSAMLEditInfo,
-    RunSAMLInfo,
-    UserCreateInfo,
-    UserInfo,
-)
+from tagoio_sdk.modules.Resources.Notification_Type import (
+    NotificationCreate, NotificationcreateReturn, NotificationInfo)
+from tagoio_sdk.modules.Resources.Run_Type import (CustomDomainCreate,
+                                                   CustomDomainInfo,
+                                                   LoginAsUserOptions,
+                                                   LoginResponse, RunInfo,
+                                                   RunSAMLEditInfo,
+                                                   RunSAMLInfo, UserCreateInfo,
+                                                   UserInfo)
 from tagoio_sdk.modules.Utils.dateParser import dateParser, dateParserList
