+++ src/tagoio_sdk/modules/Resources/Run_Type.py
@@ -1,13 +1,8 @@
 from datetime import datetime
 from typing import Dict, Literal, Optional, TypedDict, Union
 
-from tagoio_sdk.common.Common_Type import (
-    ExpireTimeOption,
-    GenericID,
-    GenericToken,
-    TagsObj,
-)
-
+from tagoio_sdk.common.Common_Type import (ExpireTimeOption, GenericID,
+                                           GenericToken, TagsObj)
 
 ThemeOption = Literal[
     "actionSchedule",
